ENH: allow fftn and ifftn to use cufftMakePlanMany
The argument plan_type was added. The default of '1d' uses the existing cufftMakePlan1d code.
If plan_type=='nd', cufftMakePlanMany is used instead. This can often be faster in practice.

Currently, cufftMakePlanMany is limited to at most 3 axes (this is a limitation in the CUFFT library).
The data to be transformed also must currently be complex and contiguous. The axes to be transformed must
be adjacent and must include either the first or last axis of the array.
The data can be either C or F contiguous.

The argument plan was also added and can be used to pass in a precomputed plan to avoid the
overhead of planning on repeated transforms.